+++ src/ts/ui/level_select.ts
@@ -51,9 +51,9 @@ export abstract class LevelSelect {
 			menu.home.show();
 		}, undefined, undefined, state.modification === 'gold');
 
-		menu.setupButton(this.prevButton, 'play/prev', () => this.cycleMission(-1), true, true, state.modification === 'gold');
+		menu.setupButton(this.prevButton, 'play/prev', () => this.cycleMission(-1), true, true, state.modification === 'gold', true);
 		menu.setupButton(this.playButton, 'play/play', () => this.playCurrentMission(), true, undefined, state.modification === 'gold');
-		menu.setupButton(this.nextButton, 'play/next', () => this.cycleMission(1), true, true, state.modification === 'gold');
+		menu.setupButton(this.nextButton, 'play/next', () => this.cycleMission(1), true, true, state.modification === 'gold', true);
 	}
 
 	abstract initProperties(): void;