+++ version_history.md
@@ -1,5 +1,10 @@
 # Version History
 
+## 2.2.2
+- Added reorderable touch action buttons
+- Added support for holding down prev/next buttons in level select
+- Fixed some styling
+
 ## 2.2.1
 - Fixed teleporting sound not playing